shaded model is now draw to the screen

+++ includes/Model.hpp
@@ -29,8 +29,8 @@ public:
 	void	Render(ID3D11DeviceContext* const pDeviceContext);
 	uint8_t	GetIndexCount() const;
 private:
-	bool	InitializeVertexData(VertexData** pVertices, uint8_t** pIndices);
-	bool	InitializeBuffers(ID3D11Device* const pDevice, VertexData const* pVertices, uint8_t const* pIndices);
+	bool	InitializeVertexData(VertexData** pVertices, uint32_t** pIndices);
+	bool	InitializeBuffers(ID3D11Device* const pDevice, VertexData const* pVertices, uint32_t const* pIndices);
 private:
 	ID3D11Buffer*	__vertexBuffer;
 	ID3D11Buffer*	__indexBuffer;
